Fixed fundamental data to use float type

+++ src/schema/fundamental.js
@@ -2,16 +2,16 @@ const { fundamental: fundamentalModel } = require('../models');
 
 const typeDefs = `
   type Fundamental {
-    open: Int
-    high: Int
-    low: Int
-    volume: Int
-    average_volume: Int
-    high_52_weeks: Int
-    low_52_weeks: Int
-    market_cap: Int
-    dividend_yield: Int
-    pe_ratio: Int
+    open: Float
+    high: Float
+    low: Float
+    volume: Float
+    average_volume: Float
+    high_52_weeks: Float
+    low_52_weeks: Float
+    market_cap: Float
+    dividend_yield: Float
+    pe_ratio: Float
     description: String
     Instrument: String
   }
